fix: overwrite old table.js file, replacing it with new table
